% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_shiny.R
\name{anova_shiny_server}
\alias{anova_shiny_server}
\title{ANOVA Shiny Server}
\usage{
anova_shiny_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
a function with Shiny server logic.
}
\description{
ANOVA Shiny Server
}
