% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_vis.R
\name{anova_vis}
\alias{anova_vis}
\title{Analysis of Variance (ANOVA) Graphic}
\usage{
anova_vis(
  Y,
  group,
  plot_datapoints = TRUE,
  plot_group_means = TRUE,
  plot_boxplot = FALSE,
  plot_group_variances = TRUE,
  plot_group_sd = TRUE,
  plot_ms_within = TRUE,
  plot_ms_between = TRUE,
  plot_between_group_variances = FALSE,
  plot_unit_line = TRUE,
  plot_grand_mean = TRUE,
  plot_sd_line = FALSE,
  plot_pooled_sd = FALSE,
  xlab = "Deviation Contrast",
  ylab = "Dependent Variable",
  grand_mean_col = "blue",
  sd_line_col = "maroon",
  pooled_sd_col = "steelblue3",
  ms_within_col = "#fdc086",
  ms_between_col = "#7fc97f",
  box_width = 0.04,
  box_color = "grey50",
  plot_group_labels = FALSE,
  ...
)
}
\arguments{
\item{Y}{the dependent variable.}

\item{group}{the independent, or grouping, variable.}

\item{plot_datapoints}{whether to plot the individual observations.}

\item{plot_group_means}{whether to plot the group means.}

\item{plot_boxplot}{whether to plot the box plots.}

\item{plot_group_variances}{whether to plot the group variances.}

\item{plot_group_sd}{whether to plot the group standard deviations.}

\item{plot_ms_within}{whether to plot the mean square within.}

\item{plot_ms_between}{whether to plot the mean square between.}

\item{plot_between_group_variances}{whether to plot the between group variances.}

\item{plot_unit_line}{whether to plot the unit line (i.e. y = x).}

\item{plot_grand_mean}{whether to plot the grand mean.}

\item{plot_sd_line}{whether to plot the standard deviation.}

\item{plot_pooled_sd}{whether to plot the pooled standard deviation.}

\item{xlab}{label for the x-axis.}

\item{ylab}{label for the y-axis.}

\item{grand_mean_col}{color for the grand mean.}

\item{sd_line_col}{color for the standard deviation line.}

\item{pooled_sd_col}{color for the pooled standard deviation.}

\item{ms_within_col}{color for the mean square within.}

\item{ms_between_col}{color for the mean square between.}

\item{box_width}{width of the box express as a percentage of the width of the x-axis.}

\item{box_color}{color of the box plots.}

\item{plot_group_labels}{whether to plot the group labels on the figure.}

\item{...}{currently unused.}
}
\value{
a ggplot2 expression.
}
\description{
This is a \code{ggplot2} adaptation of Pruzek and Helmreich's \code{granova::granova.1w}
function. This also provides parameters to customize which features are
included on the plot.
}
\examples{
data(hand_washing)
anova_vis(hand_washing$Bacterial_Counts, hand_washing$Method)
}
\references{
Pruzek & Helmreich (2010). \href{http://moderngraphics11.pbworks.com/f/ElementalGraphics4ANOVA.RP+JH.pdf}{Elemental Graphics for Analysis of Variance using the R Package granova}
}
