% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_vis.R
\name{bayes_vis}
\alias{bayes_vis}
\title{Visualization for Bayes Formula}
\usage{
bayes_vis()
}
\description{
\deqn{
P(A | B) = \frac{P(B|A) \items P(A)}{P(B)}
}
}
\details{
where

\deqn{
P(B) = P(A) \times P(B | A) + P(not A) \times P(not A | A)
}
}
\examples{
# Adapted from Tversky & Kaheman (1974) https://www.jstor.org/stable/1738360
p_a <- 1/21
p_not_a <- 1 - p_a
}
