% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plots.R
\name{beta_plot}
\alias{beta_plot}
\title{Plot a beta distribution}
\usage{
beta_plot(
  shape1,
  shape2,
  cv = 0.5,
  tail = TRUE,
  limits = c(0, 1),
  color = "steelblue"
)
}
\arguments{
\item{shape1}{non-negative parameters of the Beta distribution.}

\item{shape2}{non-negative parameters of the Beta distribution.}

\item{cv}{the critical value(s) to shade.}

\item{limits}{the limits of the plot.}

\item{color}{the color to fill in the area.}
}
\description{
Plot a beta distribution
}
