% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus_shiny.R
\name{calculus_shiny_ui}
\alias{calculus_shiny_ui}
\alias{calculus_shiny_server}
\alias{calculus_shiny}
\title{Shiny UI}
\usage{
calculus_shiny_ui()

calculus_shiny_server(input, output, session)

calculus_shiny(...)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{...}{other parameters passed to \link[shiny:shinyApp]{shiny::shinyApp}}
}
\value{
a Shiny UI object.

a function with Shiny server logic.
}
\description{
Shiny UI

Shiny server

Run the Shiny server
}
