% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus_shiny.R
\name{calculus_shiny_server}
\alias{calculus_shiny_server}
\title{Shiny server}
\usage{
calculus_shiny_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
a function with Shiny server logic.
}
\description{
Shiny server
}
