% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_shiny.R
\name{correlation_shiny_server}
\alias{correlation_shiny_server}
\title{Shiny server for maximum likelihood estimation}
\usage{
correlation_shiny_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
a function with Shiny server logic.
}
\description{
Shiny server for maximum likelihood estimation
}
