% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualStats-package.R
\docType{data}
\name{depression}
\alias{depression}
\title{Depression, anxiety and affect}
\format{
a data frame with 200 observations of 3 variables.
\describe{
\item{depression}{Depression severity}
\item{anxiety}{Anxiety severity}
\item{affect}{Positive affect}
}
}
\description{
The data set used in this set of visualizations has been provided by Professor
\href{https://www.albany.edu/psychology/faculty/james-boswell}{James Boswell},
and is from a psychological treatment study. Baseline indices of Depression severity, Anxiety
Severity, and Positive Affect were acquired on a sample of clients prior to treatment. These
simple measurements permit investigation of the relative and interactive importance of Anxiety
and Positivity in their relationship to Depression scores. It is not surprising that Anxiety
Scores and Depression scores are positively related.
}
\details{
The interesting question is whether Positive Affect moderates this relationship. The analysis
presented in these visualizations is superficial and is only intended for instructional purposes
in seeing interaction and Simple Slopes. Any strong scientific conclusions based on these
depictions would be premature.

See https://bcdudek.net/shinyapps.html for more information.
}
\keyword{datasets}
