% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus_funs.R
\name{derivative_plot}
\alias{derivative_plot}
\title{Derivative plot}
\usage{
derivative_plot(
  fun = normal_fun,
  x_value = 1,
  delta_x = 0.1,
  view_xmin = x_value - 2,
  view_xmax = x_value + 2,
  ...
)
}
\arguments{
\item{delta_x}{the difference in x used to estimate the derivative.}

\item{view_xmin}{minimum x-value for the plot view.}

\item{view_xmax}{maximum x-value for the plot view.}

\item{...}{other parameters passed to f.}

\item{f}{function used to calculate the derivative.}

\item{x}{x-coordinate where the derivative is desired.}
}
\description{
Derivative plot
}
