% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualStats-package.R
\docType{data}
\name{distributions}
\alias{distributions}
\title{Distributions}
\format{
a data frame with 18 observations of 7 variables.
\describe{
\item{label}{Distribution label.}
\item{function_name}{The name of the distribution function in R, excluding the first letter
(i.e. the \code{d}, \code{r}, \code{p}, or \code{q})}
\item{xmin}{The minimal x-value for the plotting range.}
\item{xmax}{The maximum x-value for the plotting range.}
\item{xvals}{The \code{xvals} parameter for the \code{\link[=plot_distributions]{plot_distributions()}} function. This can be
a comma separated list.}
\item{arguments}{Any other arguments for the distribution function.}
\item{Wikepedia}{Link the the Wikipedia article for the given distribution.}
}
}
\description{
This data frame provides information for all of the distribution functions in R. It is
primarily used by the \code{\link[=plot_distributions]{plot_distributions()}} function and corresponding Shiny application.
Specifically it provides the default parameters for the Shiny application.
}
\keyword{datasets}
