% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_shiny.R
\name{distributions_shiny}
\alias{distributions_shiny}
\alias{distributions_shiny_ui}
\alias{distributions_shiny_server}
\title{Run the Shiny server for visualizing distributions}
\usage{
distributions_shiny(...)

distributions_shiny_ui()

distributions_shiny_server(input, output, session)
}
\arguments{
\item{...}{other parameters passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
results of shinyApp

a Shiny UI object.

a function with Shiny server logic.
}
\description{
This Shiny application provides an interactive interface to the \code{\link[=plot_distributions]{plot_distributions()}} function.
}
