% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_errors_vis.R
\name{gaussian_errors_vis}
\alias{gaussian_errors_vis}
\title{Adapted from http://web.vu.lt/mif/a.buteikis/wp-content/uploads/PE_Book/3-4-UnivarMLE.html}
\usage{
gaussian_errors_vis(df, ...)
}
\description{
Adapted from http://web.vu.lt/mif/a.buteikis/wp-content/uploads/PE_Book/3-4-UnivarMLE.html
}
