% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linear_assumption_check.R
\name{get_numeric_vars}
\alias{get_numeric_vars}
\title{Utility function that will return numeric columns used in the regression model.}
\usage{
get_numeric_vars(glm_out)
}
\arguments{
\item{glm_out}{the results of \code{\link[stats:glm]{stats::glm()}}.}
}
\description{
Utility function that will return numeric columns used in the regression model.
}
