% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualStats-package.R
\docType{data}
\name{hand_washing}
\alias{hand_washing}
\title{Bacteria counts after various hand washing techniques.}
\format{
a data frame with 32 ovservations of 2 variables.
\describe{
\item{Bacterial_Counts}{the bacterial count after washing hands}
\item{Method}{the method of handwashing used}
}
}
\description{
From chapter 28 of De Veaux, Velleman, and Bock (2009):
}
\details{
"A student decided to investigate just how effective washing with soap is in
eliminating bacteria. To do this she tested four different methods -- washing
with water only, washing with regular soap, washing with antibacterial soap (ABS),
and spraying hands with antibacterial spray (AS) (containing 65\% ethanol as
an active ingredient). Her experiement consisted of one experimental factor,
the washing \emph{Method}, and four levels."
}
\references{
De Veaux, R.D., Velleman, P.F., & Bock, D.E. (2009). \emph{Intro Stats} (4th ed). Pearson.
}
\keyword{datasets}
