% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_sample_vis.R
\name{independent_sample_vis}
\alias{independent_sample_vis}
\title{Independent Sample Plot}
\usage{
independent_sample_vis(
  df,
  group_col = names(df)[1],
  value_col = names(df)[2],
  colors = c("#66c2a5", "#8da0cb", "#fc8d62"),
  plot_se = TRUE,
  plot_distribuition = TRUE,
  plot_zero_difference = TRUE,
  plot_ci = TRUE,
  plot_means = TRUE,
  plot_points = ifelse(nrow(df) < 60, TRUE, FALSE)
)
}
\arguments{
\item{df}{a data.frame containing the data.}

\item{group_col}{the name of the column in \code{df} containing the group membership.}

\item{value_col}{the name of the column in \code{df} containing the values.}

\item{colors}{vector of length three for the group colors. The last value is the color for the
differences.}

\item{plot_se}{whether to plot the standard error distributions (to the right of the points.}

\item{plot_distribuition}{whether to plot the raw distributions (to the left of the points).}

\item{plot_zero_difference}{whether to plot a horizontal line representing a zero difference.}

\item{plot_ci}{whether to plot confidence intervals as vertical bars.}

\item{plot_means}{whether to plot the group means.}

\item{plot_points}{whether to plot the raw observations.}
}
\value{
a ggplot2 expression.
}
\description{
This is adapted from the plots introduced by Gardner and Altman (1986) for displaying overlapping
confidence intervals for independent sample tests.
}
\examples{
set.seed(42)
# Can adjust these parameters
n <- 20
mean1 <- 10
sd1 <- 2
mean2 <- 14
sd2 <- 2
df <- data.frame(
	group = c(rep('Group A', n / 2),
			  rep('Group B', n / 2)),
	value = c(rnorm(n = (n / 2), mean = mean1, sd = sd1),
			  rnorm(n = (n / 2), mean = mean2, sd = sd2))
)

independent_sample_vis(df)
}
\references{
Gardner, M.J. & Altman, D.G. (1986). Confidence intervals rather than P values:
estimation rather than hypothesis testing. \emph{British Medical Journal, 292}(6522) 746-750.
doi:10.1136/bmj.292.6522.746
}
