% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus_funs.R
\name{integral_plot}
\alias{integral_plot}
\title{Integral plot using Riemann sums.}
\usage{
integral_plot(
  fun = normal_fun,
  xmin = 0,
  xmax = 1,
  view_xmin = xmin - 0.5 * (xmax - xmin),
  view_xmax = xmax + 0.5 * (xmax - xmin),
  n = 10,
  rect_alpha = 0.5,
  rect_color = "black",
  rect_fill = "grey50"
)
}
\arguments{
\item{fun}{function used to calculate Riemann sums.}

\item{view_xmin}{minimum x-value for the plot view.}

\item{view_xmax}{maximum x-value for the plot view.}

\item{n}{number of rectangles used to estimate Riemann sum.}

\item{min}{minimum x-value.}

\item{max}{maximum x-value.}
}
\description{
Integral plot using Riemann sums.
}
