% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBinary.R
\name{isBinary}
\alias{isBinary}
\title{Check if a vector is dichotomous.}
\usage{
isBinary(x)
}
\arguments{
\item{x}{vector to check.}
}
\value{
TRUE if the vector is binary.
}
\description{
This function checks to see if a vector is dichotomous. This includes
logical vectors and numeric vectors that contain only 0s and 1s.
}
