% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_by.R
\name{kurtosi}
\alias{kurtosi}
\title{Copied from the psych package}
\usage{
kurtosi(x, na.rm = TRUE, type = 3)
}
\arguments{
\item{x}{numeric vector.}

\item{na.rm}{whether to remove missing values.}

\item{type}{see \code{\link[psych:kurtosi]{psych::kurtosi()}}.}
}
\description{
Copied from the psych package
}
