% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_shiny.R
\name{loess_shiny}
\alias{loess_shiny}
\title{Run the Loess Shiny server}
\usage{
loess_shiny(...)
}
\arguments{
\item{...}{other parameters passed to \link[shiny:shinyApp]{shiny::shinyApp}}
}
\description{
Run the Loess Shiny server
}
