% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_vis.R
\name{loess_vis}
\alias{loess_vis}
\title{Loess Visualization}
\usage{
loess_vis(
  formula,
  data,
  center = median(data[, all.vars(formula)[2], drop = TRUE]),
  degree = 2,
  span = 0.3,
  draw_loess = FALSE,
  show_loess = TRUE,
  xlab = all.vars(formula)[2],
  ylab = all.vars(formula)[1],
  ...
)
}
\arguments{
\item{formula}{regression formula.}

\item{data}{data frame containing the data.}

\item{center}{where to display the current Loess estimation.}

\item{degree}{whether to estimate local linear (1) or quadratic (2) regressions.}

\item{span}{the parameter which controls the degree of smoothing.}

\item{draw_loess}{whether to draw the Loess regression line up to \code{center}.}

\item{show_loess}{whether to draw the complete Loess regression line.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{...}{currently not used.}
}
\description{
Loess Visualization
}
\examples{
data("faithful")
loess_vis(formula = eruptions ~ waiting, data = faithful)
}
