% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{Logistic Function}
\usage{
logistic(x, beta0, beta1)
}
\arguments{
\item{x}{vector of values.}

\item{beta0}{intercept.}

\item{beta1}{slope.}
}
\description{
Logistic Function
}
