% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_binomial.R
\name{loglikelihood_binomial}
\alias{loglikelihood_binomial}
\title{Calculate the log likelihood for a binomial distribution.}
\usage{
loglikelihood_binomial(parameters, predictor, outcome)
}
\arguments{
\item{parameters}{the regression parameters. Element one is the intercept,
element two is the slope or beta coefficient.}

\item{predictor}{vector of predictors.}

\item{outcome}{vector of outcomes.}
}
\value{
the log likelihood.
}
\description{
Calculate the log likelihood for a binomial distribution.
}
