% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_normal.R
\name{loglikelihood_normal}
\alias{loglikelihood_normal}
\title{Calculate the log likelihood for a normally distributed dependent variable.}
\usage{
loglikelihood_normal(parameters, predictor, outcome)
}
\arguments{
\item{parameters}{the regression parameters. Element one is the intercept,
element two is the slope, and element three is the error (root mean
square error).}

\item{predictor}{vector of predictors.}

\item{outcome}{vector of outcomes.}
}
\value{
the log likelihood.
}
\description{
Calculate the log likelihood for a normally distributed dependent variable.
}
