% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualStats-package.R
\docType{data}
\name{manufacturing}
\alias{manufacturing}
\title{Product strength predicted from processing time, temperature, and pressure}
\format{
a data frame with 29 observations of 4 variables
\describe{
\item{Strength}{}
\item{Temperature}{}
\item{Pressure}{}
\item{Time}{}
}
}
\description{
This data set is designed to explore the interactions effects. The independent variables
(processing time, temperature, and pressure) affect the dependent variable (product strength).
There is an interaction between \code{Temperature} and \code{Pressure}.
}
\references{
Frost, J. (2020). \emph{Regression Analysis: An Intuitive Guide for Using and Interpreting
Linear Models} Statistics By Jim Publishing.
https://statisticsbyjim.com/regression/interaction-effects/
}
