% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_shiny.R
\name{mle_shiny}
\alias{mle_shiny}
\title{Visualizing Maximum Likelihood Estimation Shiny Application}
\usage{
mle_shiny(df, default_x, default_y, ...)
}
\arguments{
\item{df}{a data.frame to explore.}

\item{default_x}{default variable for the x-axis.}

\item{default_y}{default variable for the y-axis.}

\item{...}{other parameters passed to \link[shiny:shinyApp]{shiny::shinyApp}}
}
\description{
This will start a shiny app explore maximum likelihood estimation visually.
}
\references{
http://rstudio.com/shiny
}
