% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_shiny.R
\name{mle_shiny_server}
\alias{mle_shiny_server}
\title{Shiny server for maximum likelihood estimation}
\usage{
mle_shiny_server(input, output, session)
}
\value{
a function with Shiny server logic.
}
\description{
Shiny server for maximum likelihood estimation
}
