% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_regression_shiny.R
\name{multiple_regression_shiny}
\alias{multiple_regression_shiny}
\title{Run the Shiny server for visualizing multiple regression}
\usage{
multiple_regression_shiny(...)
}
\arguments{
\item{...}{other parameters passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\value{
results of shinyApp
}
\description{
This Shiny application provides an interactive interface to the \code{\link[=multiple_regression_vis]{multiple_regression_vis()}} function.
}
