% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_regression_vis.R
\name{multiple_regression_vis}
\alias{multiple_regression_vis}
\title{3d plot scatter plot with regression plane}
\usage{
multiple_regression_vis(
  y,
  x1,
  x2,
  interaction = FALSE,
  plot_regression = TRUE,
  plot_residuals = TRUE,
  plot_slopes = NULL,
  y_lab = "Y",
  x1_lab = "X1",
  x2_lab = "X2"
)
}
\arguments{
\item{y}{dependent varaible.}

\item{x1}{independent variable.}

\item{x2}{independent variable.}

\item{interaction}{whether to plot the plane including the interaction effects (i.e. warped plane).}

\item{plot_regression}{whether to plot the regression plane.}

\item{plot_residuals}{if \code{TRUE} the residuals will be plotted as line segments.}

\item{plot_slopes}{either \code{x1} or \code{x2} to plot the simple slopes (i.e. plus and minus one
standard deviation) or \code{NULL} to exclude.}

\item{y_lab}{label for the y-axis.}

\item{x1_lab}{label for x1-axis.}

\item{x2_lab}{label for x2-axis.}
}
\value{
a plot_ly expression.
}
\description{
3d plot scatter plot with regression plane
}
\examples{
data(depression, package = 'VisualStats')
multiple_regression_vis(y = depression$depression,
    x1 = depression$anxiety,
    x2 = depression$affect,
    y_lab = 'Depression',
    x1_lab = 'Anxiety',
    x2_lab = 'Affect',
    interaction = TRUE)

}
