% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus_funs.R
\name{normal_fun}
\alias{normal_fun}
\title{Normal distribution function}
\usage{
normal_fun(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{x value}
}
\value{
y value for the given x.
}
\description{
Normal distribution function
}
