% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_save.R
\name{optim_save}
\alias{optim_save}
\title{Warpper to optim to save iterations.}
\usage{
optim_save(par, fn, ...)
}
\arguments{
\item{par}{initial parameters to to be optimized over.}

\item{fn}{the function to minimized (or maximized).}

\item{...}{other parameters passed to optim.}
}
\value{
the results of optim with two additional elements, iterations with a
a list of the values at each iteration of the algorithm and
iterations_df which is a data.frame version of the list.
}
\description{
This function wraps the \link{stats::optim} function and saves the parameters
and likelihood estimation at each step of the algorithm.
}
\seealso{
stats::optim
}
