% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Plot the four distributions functions.}
\usage{
plot_distributions(
  dist,
  xvals,
  xmin,
  xmax,
  args = list(),
  palette = c(d = "#1b9e77", r = "#d95f02", p = "#7570b3", q = "#e7298a")
)
}
\description{
Plot the four distributions functions.
}
