% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inner_circle.R
\name{plot_inner_circle}
\alias{plot_inner_circle}
\title{Plots a circle within a circle.}
\usage{
plot_inner_circle(percent = 0.5, y_inner = c("center", "bottom"))
}
\description{
Plots a circle within a circle.
}
