% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inner_square.R
\name{plot_inner_square}
\alias{plot_inner_square}
\title{Plots a circle within a circle.}
\usage{
plot_inner_square(percent = 0.5, inner = c("center", "bottom"))
}
\description{
Plots a circle within a circle.
}
