% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_likelihood.R
\name{plot_likelihood}
\alias{plot_likelihood}
\title{Scatter plot with likelihood superimposed.}
\usage{
plot_likelihood(
  x,
  y,
  height = 0.4,
  k = 4,
  pt = 1,
  intercept = lm(y ~ x)$coefficients[1],
  slope = lm(y ~ x)$coefficients[2],
  sigma = sqrt(sum(resid(lm(y ~ x))^2)/length(x)),
  ...
)
}
\arguments{
\item{x}{the predictor}

\item{y}{the outcome}

\item{height}{the proportion of the x-axis the distribution should occupy.
That is, a height of 0.4 means the width of the distribution will
occupy approximately 40 percent of the plot width.}

\item{k}{the number of standard deviations to plot}

\item{pt}{the point to highlight and draw the likelihood of.}

\item{sigma}{the root mean square error}
}
\description{
Scatter plot with likelihood superimposed.
}
