% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linear_assumption_check.R
\name{plot_linear_assumption_check}
\alias{plot_linear_assumption_check}
\title{Visual plot to check the assumption that the predicted values are linearly
related to the predictors.}
\usage{
plot_linear_assumption_check(
  glm_out,
  n_groups,
  vars = get_numeric_vars(glm_out),
  na.rm = TRUE,
  x_lab = "Predicted Probabilities",
  y_lab = all.vars(glm_out$formula)[1],
  x_limits = c(0, 1),
  coord_equal = TRUE,
  coord_flip = FALSE,
  widths = NULL,
  legend.position = "bottom",
  varplot_nrow = NULL,
  varplot_ncol = NULL
)
}
\arguments{
\item{glm_out}{the results of the logistic regression (i.e. the result from
\code{\link{glm}}).}

\item{vars}{character vector with the name of the independent variables to
include. If omitted, all numeric variables will be used.}

\item{widths}{the relative widths of the two plots. If you only want the first
plot, use \code{widths = c(0, 1)}; if only the second plot, use
\code{widths = c(1, 0)}.}

\item{the}{number of groups to use. By default groups will have a minimum of
six observations up to a maximum of 10 groups.}
}
\value{
a ggplot2 expression.
}
\description{
Visual plot to check the assumption that the predicted values are linearly
related to the predictors.
}
\examples{
study <- data.frame(
    Hours=c(0.50,0.75,1.00,1.25,1.50,1.75,1.75,2.00,2.25,2.50,2.75,3.00,
		       3.25,3.50,4.00,4.25,4.50,4.75,5.00,5.50),
    Pass=c(0,0,0,0,0,0,1,0,1,0,1,0,1,0,1,1,1,1,1,1)
)
lr.out <- glm(Pass ~ Hours, data = study,
			  family = binomial(link='logit'))
plot_linear_assumption_check(glm_out = lr.out, n_groups = 5)
}
