% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_sample_vis.R
\name{plot_p_value}
\alias{plot_p_value}
\title{Internal function for printing p-values.}
\usage{
plot_p_value(p, threshold = 0.01, ...)
}
\arguments{
\item{p}{the p-value.}

\item{threshold}{what threshold should the p-value be printed as p < threshold.}

\item{...}{other parameters passed to \code{\link[=print_num]{print_num()}}.}
}
\value{
a character representation of the p-value.
}
\description{
If the p-value is less than the threshold, then p < X will be printed, otherwise
\code{print_num()} will be called.
}
