% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualStats-package.R
\docType{data}
\name{poverty}
\alias{poverty}
\title{Various characteristics of States (and District of Columbia) including poverty}
\format{
a data frame with 51 observations of 6 variables
\describe{
\item{state}{The state.}
\item{metro_res}{Percent of people in the state living in a metropolitan area.}
\item{white}{Percent of the state population that is white.}
\item{hs_grad}{Percent of the state population that has a high school diploma.}
\item{poverty}{Percent of the state population living in poverty.}
\item{female_house}{Percent of the state where the female is head of household.}
}
}
\description{
This dataset was originally used by Gelman and Hill (2007) to demonstrate multiple regression.
In particular multicollinearity predicting \code{poverty} from \code{female_house} and \code{white}.
}
\references{
Gelman, H. & Hill, J. (2007). \emph{Data Analysis using Regression and Multilevel/Hierarchical Models.}
Cambridge University Press.
}
