% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_sample_vis.R
\name{print_num}
\alias{print_num}
\title{Internal function for printing numbers.}
\usage{
print_num(x, big.mark = ",", scientific = FALSE, digits = 3, ...)
}
\description{
Internal function for printing numbers.
}
