% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_sample_vis.R
\name{print_p_value}
\alias{print_p_value}
\title{Internal function for printing p-values.
If the p-value is less than the threshold, then p < X will be printed, otherwise
\code{\link[=print_num]{print_num()}} will be callsed.}
\usage{
print_p_value(p, threshold = 0.01, ...)
}
\description{
Internal function for printing p-values.
If the p-value is less than the threshold, then p < X will be printed, otherwise
\code{\link[=print_num]{print_num()}} will be callsed.
}
