% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_shiny.R
\name{r_squared_shiny}
\alias{r_squared_shiny}
\title{Run the Shiny server for visualizing R-Squared}
\usage{
r_squared_shiny(...)
}
\arguments{
\item{...}{other parameters passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\value{
results of shinyApp
}
\description{
Run the Shiny server for visualizing R-Squared
}
