% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_shiny.R
\name{r_squared_shiny_server}
\alias{r_squared_shiny_server}
\title{R-Squared Shiny Server}
\usage{
r_squared_shiny_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
a function with Shiny server logic.
}
\description{
R-Squared Shiny Server
}
