% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_shiny.R
\name{regression_shiny}
\alias{regression_shiny}
\alias{regression_shiny_ui}
\alias{regression_shiny_server}
\title{Shiny application for visualization of correlations, cross products, regression, and residuals.}
\usage{
regression_shiny(...)

regression_shiny_ui()

regression_shiny_server(input, output, session)
}
\arguments{
\item{...}{other parameters passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
a Shiny UI object.

a function with Shiny server logic.
}
\description{
Shiny application for visualization of correlations, cross products, regression, and residuals.

Shiny UI for correlation

Shiny server for maximum likelihood estimation
}
