% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_vis.R
\name{regression_vis}
\alias{regression_vis}
\title{Correlation and cross product visualization}
\usage{
regression_vis(
  df,
  x_var = names(df)[1],
  y_var = names(df)[2],
  plot_x_mean = TRUE,
  plot_y_mean = TRUE,
  plot_regression = FALSE,
  regression_line_color = "grey30",
  plot_x_deviations = NULL,
  plot_y_deviations = NULL,
  x_deviation_color = "darkblue",
  y_deviation_color = "darkblue",
  deviation_size = 1.5,
  plot_positive_cross_products = FALSE,
  plot_negative_cross_products = FALSE,
  plot_cross_products = NULL,
  positive_cross_product_fill = "lightblue",
  negative_cross_product_fill = "darkred",
  positive_cross_product_color = cross_product_color,
  negative_cross_product_color = cross_product_color,
  cross_product_color = "grey30",
  cross_product_fill = "#F5C710",
  cross_product_alpha = 0.1,
  plot_residuals = NULL,
  plot_residuals_squared = NULL,
  residuals_squared_color = "grey30",
  residuals_squared_fill = "green",
  residuals_squared_alpha = 0.1,
  residual_color = "darkred",
  residual_size = 1.5
)
}
\arguments{
\item{df}{data.frame with the data to plot.}

\item{x_var}{the name of the variable for the x-axis.}

\item{y_var}{the name of the variable fro teh y-axis.}

\item{plot_x_mean}{plot a line for the mean of x.}

\item{plot_y_mean}{plot a line for the mean of y.}

\item{plot_regression}{whether to plot the linear regression line.}

\item{regression_line_color}{color for the regression line.}

\item{plot_x_deviations}{either a logical or numerical vector for the rows in \code{df}
to plot the x deviations.}

\item{plot_y_deviations}{either a logical or numerical vector for the rows in \code{df}
to plot the y deviations.}

\item{x_deviation_color}{the color of the line for x deviations.}

\item{y_deviation_color}{the color of the line for y deviations.}

\item{deviation_size}{of the line when plotting deviations.}

\item{plot_positive_cross_products}{plot all the cross products in the 1st and 3rd quadrants.}

\item{plot_negative_cross_products}{plot all the cross products in the 2nd and 4th quandrants.}

\item{plot_cross_products}{either a logicial or numerical vector for the rows in \code{df}
to plot the cross products.}

\item{positive_cross_product_fill}{the fill for the positive cross products.}

\item{negative_cross_product_fill}{the fill for the negative cross products.}

\item{positive_cross_product_color}{the color of the square for positive cross products.}

\item{negative_cross_product_color}{the color of the square for negative cross products.}

\item{cross_product_color}{the color of the square for cross products.}

\item{cross_product_fill}{the fill for the cross products plotted by \code{plot_cross_products}.}

\item{cross_product_alpha}{the alpha level (transparency) for the cross product fill.}

\item{plot_residuals}{whether to plot the residuals.}

\item{plot_residuals_squared}{whether to plot the squared residuals.}

\item{residuals_squared_color}{color for the residuals.}

\item{residuals_squared_fill}{fill color for the squared residuals.}

\item{residuals_squared_alpha}{alpha (transparency) level for the squared residuals.}

\item{residual_color}{color for the residuals.}

\item{residual_size}{size of the residuals.}
}
\value{
a ggplot2 expression.
}
\description{
Correlation and cross product visualization
}
\examples{
df <- mtcars[,c('wt', 'mpg')]
cross_products <- abs(df[,1] * df[,2])
cross_products == max(cross_products) # Find the largest cross product

# Scatter plot with arrows showing the largest cross product
regression_vis(df,
				plot_x_mean = TRUE,
				plot_y_mean = TRUE,
				plot_positive_cross_products = FALSE,
				plot_negative_cross_products = FALSE,
				plot_x_deviations = cross_products == max(cross_products),
				plot_y_deviations = cross_products == max(cross_products))

# Scatter plot with the largest cross product as a rectangl
regression_vis(df,
				plot_x_mean = TRUE,
				plot_y_mean = TRUE,
				plot_positive_cross_products = FALSE,
				plot_negative_cross_products = FALSE,
				plot_cross_products = cross_products == max(cross_products),
				cross_product_alpha = 0.5)

# Scatter plot with all the cross products.
regression_vis(df,
				plot_x_mean = TRUE,
				plot_y_mean = TRUE,
				plot_positive_cross_products = TRUE,
				plot_negative_cross_products = TRUE)

}
