% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderRmd.R
\name{renderRmd}
\alias{renderRmd}
\title{Render an RMarkdown file to be included in a Shiny app.}
\usage{
renderRmd(path, input, envir = new.env(), quiet = TRUE)
}
\description{
Render an RMarkdown file to be included in a Shiny app.
}
