% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus_funs.R
\name{riemann}
\alias{riemann}
\title{Calculate Riemann sums}
\usage{
riemann(fun, min, max, n = 2, ...)
}
\arguments{
\item{fun}{function used to calculate the intergral using Riemann sums.}

\item{min}{minimum x-value.}

\item{max}{maximum x-value.}

\item{n}{number of rectangles used to estimate Riemann sum.}

\item{...}{other parameters passed to f.}
}
\value{
data.frame with three variables: xmin, height (i.e. f(xmin)), and area.
}
\description{
Calculate Riemann sums
}
