% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculate standard error}
\usage{
se(x)
}
\arguments{
\item{x}{a numeric vector to calculate the standard error for.}
}
\value{
the standard error.
}
\description{
Calculates the standard error of the mean using the following formula:
}
\details{
\deqn{se = \frac{s}{\sqrt{n}}}

where \emph{s} is the sample standard deviation and \emph{n} is the sample size.
}
