% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mle.R
\name{shiny_mle}
\alias{shiny_mle}
\title{Visualizing Maximum Likelihood Estimation Shiny Application}
\usage{
shiny_mle(df, default_x, default_y, ...)
}
\description{
This will start a shiny app explore maximum likelihood estimation visually.
}
\references{
http://rstudio.com/shiny
}
