% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_vis.R
\name{simulate}
\alias{simulate}
\title{Simulate a data frame for multiple regression.}
\usage{
simulate(n = 100, beta = c(5, 3, -2), r_squared = 0.8)
}
\arguments{
\item{n}{number of observations.}

\item{beta}{beta coefficients. Position one is the intercept, the rest are predictors.}

\item{r_squared}{desired R-squared value.}
}
\value{
a data.frame.
}
\description{
Simulate a data frame for multiple regression.
}
\examples{
get_r_squared <- function(desired) {
    model <- lm(y ~ x1 + x2, data = simulate(r_squared = desired, n = 100))
    return(summary(model)$r.squared)
}
df <- data.frame(desired_r_squared = seq(from = 0.05, to = 0.95, by = 0.05))
df$actual_r_squared <- sapply(df$desired_r_squared, FUN = get_r_squared)
plot(df)
abline(a=0, b=1, col="red", lty=2)
}
