% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_by.R
\name{skew}
\alias{skew}
\title{Copied from the psych package.}
\usage{
skew(x, na.rm = TRUE, type = 3)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{whether missing values shoujld be removed.}

\item{type}{see \code{\link[psych:skew]{psych::skew()}}}
}
\description{
Copied from the psych package.
}
