% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_shiny.R
\name{variance_shiny_server}
\alias{variance_shiny_server}
\title{UI function for variance application}
\usage{
variance_shiny_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object}
}
\description{
UI function for variance application
}
