% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_vis.R
\name{variance_vis}
\alias{variance_vis}
\title{Variance Visualization}
\usage{
variance_vis(
  x,
  plot_mean = TRUE,
  plot_deviances = TRUE,
  plot_deviances_x = FALSE,
  plot_deviances_y = FALSE,
  plot_population_variance = TRUE,
  plot_sample_variance = FALSE,
  plot_population_sd = FALSE,
  plot_sample_sd = FALSE,
  point_size = 3,
  variance_position = "top",
  deviation_col = "blue",
  sample_variance_col = "green",
  population_variance_col = "orange",
  sample_sd_col = "darkgreen",
  population_sd_col = "darkorange",
  sd_size = 1.5,
  xlab = "",
  ...
)
}
\arguments{
\item{x}{a vector of values to display the variance of.}

\item{plot_mean}{whether to plot the mean.}

\item{plot_deviances, plot_deviances_x, plot_deviances_y}{either a logical (TRUE to include all cross
products) or an integer vector indicating which cross products from
\code{x} are to be plotted.}

\item{plot_population_variance}{whether to plot the population variance.}

\item{plot_sample_variance}{whether to plot the sample variance.}

\item{plot_population_sd}{whether to plot the population standard deviation.}

\item{plot_sample_sd}{whether to plot the sample standard deviation.}

\item{point_size}{the size of points.}

\item{variance_position}{where to plot the sample and/or population
variances. Values can be 'top', 'bottom', or 'middle'.}

\item{deviation_col}{color of the deviations.}

\item{sample_variance_col}{color of the sample variance.}

\item{population_variance_col}{color of the population variance.}

\item{sample_sd_col}{color of the sample standard deviation.}

\item{population_sd_col}{color of the population standard deviation.}

\item{sd_size}{size of the standard deviation error bars.}

\item{xlab}{label for the x-axis.}

\item{...}{currently not used.}
}
\value{
a ggplot2 expression.
}
\description{
This function is adapted from Bruce Dudek's Shiny app available here:
https://shiny.rit.albany.edu/stat/visualizess/
}
\examples{
	x <- c(97.88, 107.91, 88.26, 115.21, 87.38)
	variance_vis(x)
}
\author{
Jason Bryer

Bruce Dudek
}
