% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualStats.R
\docType{package}
\name{VisualStats-package}
\alias{VisualStats-package}
\title{Visualizations for Statistical Tests}
\description{
Visualizations for Statistical Tests
}
\author{
\email{jason@bryer.org}
}
\keyword{shiny}
\keyword{statistics}
\keyword{visualizations}
