% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualStats-package.R
\name{vs_palette_qual}
\alias{vs_palette_qual}
\title{Color palette used for visualizations (qualitative)}
\usage{
vs_palette_qual
}
\description{
Color palette used for visualizations (qualitative)
}
