% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualStats-package.R
\name{vs_palette_seq_blues}
\alias{vs_palette_seq_blues}
\title{Color palette used for visualizations (sequential and blue)}
\usage{
vs_palette_seq_blues
}
\description{
Color palette used for visualizations (sequential and blue)
}
