% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualStats-package.R
\name{vs_palette_seq_greens}
\alias{vs_palette_seq_greens}
\title{Color palette used for visualizations (sequential and greens)}
\usage{
vs_palette_seq_greens
}
\description{
Color palette used for visualizations (sequential and greens)
}
