% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus_shiny.R
\name{calculus_shiny}
\alias{calculus_shiny}
\title{Run the Shiny server}
\usage{
calculus_shiny(...)
}
\arguments{
\item{...}{other parameters passed to \link[shiny:shinyApp]{shiny::shinyApp}}
}
\description{
Run the Shiny server
}
