% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_shiny.R
\name{correlation_shiny}
\alias{correlation_shiny}
\title{Shiny application for visualization of correlations and cross products.}
\usage{
correlation_shiny(...)
}
\arguments{
\item{...}{other parameters passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\description{
Shiny application for visualization of correlations and cross products.
}
