% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent_sample_shiny.R
\name{dependent_sample_shiny}
\alias{dependent_sample_shiny}
\title{Run the Shiny server for visualizing R-Squared}
\usage{
dependent_sample_shiny(...)
}
\arguments{
\item{...}{other parameters passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\value{
results of shinyApp
}
\description{
Run the Shiny server for visualizing R-Squared
}
