% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_by.R
\name{describe_by}
\alias{describe_by}
\title{Basic summary statistics by group}
\usage{
describe_by(x, group = NULL, type = 3, trim = 0.1, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a numeric vector to summarize.}

\item{group}{a grouping vector.}

\item{type}{type of skewness, see \link[psych:skew]{psych::skew} for details.}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each end
of x before the mean is computed. Values of trim outside that range are
taken as the nearest endpoint.}

\item{na.rm}{a logical evaluating to TRUE or FALSE indicating whether NA values
should be stripped before the computation proceeds.}

\item{...}{not currently used. Included to ensure compatibility.}

\item{mat}{not used, provided to provide direct compatability with \link[psych:describe.by]{psych::describeBy}.}
}
\value{
a data.frame with the summary statistic.
}
\description{
This is a re-implementation of the \link[psych:describe.by]{psych::describeBy} function using the \code{dplyr}
package for creating the resulting table. The goal is for this function to be
a drop-in replacement, however the results are always returned as a data.frame
consistent with the \code{mat = TRUE} parameter setting.
}
\details{
The \code{psych} package is an incredibly useful package but some of the dependencies
are problematic for deploying Shiny applications using \code{shinylive}. Since this
package only uses the \code{describeBy} function from the \code{psych} package, we
re-implemented it in order to remove the dependency on the \code{psych} package.
}
