% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_sample_vis.R
\name{independent_sample_vis}
\alias{independent_sample_vis}
\title{Independent Sample Plot}
\usage{
independent_sample_vis(
  df,
  group_col = names(df)[1],
  value_col = names(df)[2],
  colors = c("#66c2a5", "#8da0cb", "#fc8d62")
)
}
\arguments{
\item{df}{a data.frame containing the data.}

\item{group_col}{the name of the column in \code{df} containing the group membership.}

\item{value_col}{the name of the column in \code{df} containing the values.}

\item{colors}{vector of length three for the group colors. The last value is the color for the
differences.}
}
\value{
a ggplot2 expression.
}
\description{
This is adapted from the plots introduced by Gardner and Altman (1986) for displaying overlapping
confidence intervals for indepdent sample tests.
}
\references{
Gardner, M.J. & Altman, D.G. (1986). Confidence intervals rather than P values:
estimation rather than hypothesis testing. \emph{British Medical Journal, 292}(6522) 746-750.
doi:10.1136/bmj.292.6522.746
}
