% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_vis.R
\name{loess_vis}
\alias{loess_vis}
\title{Loess Visualization}
\usage{
loess_vis(
  formula,
  data,
  center = median(data[, all.vars(formula)[2], drop = TRUE]),
  degree = 2,
  span = 0.3,
  draw_loess = FALSE,
  show_loess = TRUE,
  xlab = all.vars(formula)[2],
  ylab = all.vars(formula)[1],
  ...
)
}
\description{
Loess Visualization
}
\examples{
data("faithful")
loess_vis(formula = eruptions ~ waiting, data = faithful)
}
