% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_shiny.R
\name{mle_shiny}
\alias{mle_shiny}
\title{Visualizing Maximum Likelihood Estimation Shiny Application}
\usage{
mle_shiny(df, default_x, default_y, ...)
}
\description{
This will start a shiny app explore maximum likelihood estimation visually.
}
\references{
http://rstudio.com/shiny
}
