% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Plot the four distributions functions.}
\usage{
plot_distributions(
  dist,
  xvals,
  xmin,
  xmax,
  args = list(),
  palette = c(d = "#1b9e77", r = "#d95f02", p = "#7570b3", q = "#e7298a")
)
}
\arguments{
\item{dist}{name of the distribution function to plot (e.g. \code{norm}, \code{t}).}

\item{xvals}{inputs to \code{dist} function.}

\item{xmin}{minimum x-axis value to plot.}

\item{xmax}{maximum x-axis value to plot.}

\item{args}{parameters to pass to \code{dist}.}

\item{palette}{colors to use for the four subplots. Must be a named vector with d, r, p, and q.}
}
\description{
When working with distributions in R, each distribution has four functions, namely:
}
\details{
\itemize{
\item \code{dXXX} - density function.
\item \code{rXXX} - generate random number from this distribution.
\item \code{pXXX} - returns the area to the left of the given value.
\item \code{qXXX} - returns the quantile for the given value.
}

Where \code{XXX} is the distribution name (e.g. \code{norm}, \code{binom}, \code{t}, etc.).
}
\examples{
# Normal distribution with mean 100 and standard deviation 15
plot_distributions(dist = 'norm',
                   xvals = c(80, 90, 105),
                   xmin = 55,
                   xmax = 145,
                   args = list(mean = 100, sd = 15))

# Binomial distribution
plot_distributions(dist = 'binom',
				      xvals = c(1, 3),
				      xmin = 0,
				      xmax = 10,
				      args = list(size = 10, prob = 0.35))
# Geometric distribution
plot_distributions(dist = 'geom',
                   xvals = c(1, 3),
                   args = list(prob = 0.2),
                   xmin = 0,
                   xmax = 10)

# Poisson distribution
plot_distributions(dist = 'pois',
                   xvals = c(1, 3),
                   args = list(lambda = 1),
                   xmin = 0,
                   xmax = 10)
}
