% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_vis.R
\name{plot_tiles}
\alias{plot_tiles}
\title{Tile plot}
\usage{
plot_tiles(
  p,
  percent,
  xmin,
  ymin,
  xmax,
  ymax,
  rev = FALSE,
  color = "black",
  fill = "#F5C710",
  alpha = 0.5
)
}
\description{
Tile plot
}
