% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_vis.R
\name{r_squared_vis}
\alias{r_squared_vis}
\title{Regression Sum of Squared Visualization}
\usage{
r_squared_vis(
  df,
  formu,
  x_lab = "Observed Value",
  y_lab = "Predicted Value",
  plot_total_variance = TRUE,
  plot_regression_variance = TRUE,
  plot_error_variance = FALSE,
  plot_unit_line = TRUE,
  plot_points = TRUE,
  plot_means = TRUE,
  plot_residuals = TRUE,
  plot_residuals_squared = TRUE
)
}
\arguments{
\item{df}{data.frame with the data to plot.}

\item{formu}{formula for the regression.}
}
\description{
Regression Sum of Squared Visualization
}
