% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_shiny.R
\name{variance_shiny}
\alias{variance_shiny}
\title{Visualizing Variance Shiny Application}
\usage{
variance_shiny(...)
}
\arguments{
\item{...}{other parameters passed to \link[shiny:shinyApp]{shiny::shinyApp}}
}
\description{
This will start a shiny app explore variance visually.
}
\references{
http://rstudio.com/shiny
}
