% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualStats-package.R
\docType{package}
\name{VisualStats-package}
\alias{VisualStats}
\alias{VisualStats-package}
\title{Visualizations for Statistical Analysis}
\description{
Visualizations for Statistical Analysis
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jbryer.github.io/VisualStats/}
  \item Report bugs at \url{https://github.com/jbryer/VisualStats/issues}
}

}
\author{
\email{jason@bryer.org}
}
\keyword{shiny}
\keyword{statistics}
\keyword{visualizations}
