% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualStats-package.R
\docType{data}
\name{vs_palette_qual}
\alias{vs_palette_qual}
\title{Color palette used for visualizations (qualitative)}
\format{
An object of class \code{character} of length 9.
}
\usage{
vs_palette_qual
}
\description{
Color palette used for visualizations (qualitative)
}
\keyword{datasets}
