% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualStats-package.R
\docType{data}
\name{vs_palette_seq_purples}
\alias{vs_palette_seq_purples}
\title{Color palette used for visualizations (sequential and purples)}
\format{
An object of class \code{character} of length 5.
}
\usage{
vs_palette_seq_purples
}
\description{
Color palette used for visualizations (sequential and purples)
}
\keyword{datasets}
